fix(sources): CI — extrait le cœur promotion vers app/ (fix boot Docker) + ruff format
Deux échecs CI sur la PR #1000 :
- build : ModuleNotFoundError 'scripts' — l'image Docker n'embarque que app/,
  donc `from scripts...` dans app/jobs/promote_sources_job.py plantait le
  conteneur au boot (et aurait planté Railway).
- lint : app/jobs/promote_sources_job.py non `ruff format`-clean.

Fix :
- Nouveau module app/services/source_promotion.py : gate (bias/fiabilité/
  volume/denylist), SourceMeta/Promotion, compute_promotions, load_metas,
  write_promotions. Le job importe d'ici (app -> app).
- scripts/retag_and_promote_sources.py ré-exporte ces symboles (compat CLI +
  tests inchangés) ; dépendance script -> app (bon sens).
- ruff format appliqué.
- tests/services/test_source_promotion.py : gate first-class + garde-fou
  anti-régression (le job ne réimporte jamais `scripts`).

Vérifié : suite complète 2474 passed ; import scheduler OK avec `scripts` bloqué
(simulation Docker) ; ruff check + format app/ clean.
